create profiles controller, move token to account controller
